clean up cli tool
