+++ packages/create-gator-app/src/lib/choices/framework.ts
@@ -1,0 +1,7 @@
+export const FRAMEWORKS_OPTIONS = [
+    {
+      name: "NextJS",
+      value: "nextjs",
+    },
+  ];
+
