+++ packages/create-gator-app/src/lib/choices/templates.ts
@@ -1,0 +1,7 @@
+export const TEMPLATE_OPTIONS = [
+  {
+    name: "Basic Delegator app with NextJS",
+    value: "starter",
+    description: "A basic Next.js template with App Router",
+  },
+];
