+++ packages/create-gator-app/src/lib/choices/web3auth-network.ts
@@ -1,0 +1,10 @@
+export const WEB3AUTH_NETWORK_CHOICES = [
+  {
+    name: "Sapphire Devnet",
+    value: "sapphire_devnet",
+  },
+  {
+    name: "Sapphire Mainnet",
+    value: "sapphire_mainnet",
+  },
+];
