--- packages/create-gator-app/src/lib/frameworks.ts
@@ -1,6 +1,0 @@
-export const FRAMEWORKS = [
-  {
-    name: "NextJS",
-    value: "nextjs",
-  },
-];
