+++ packages/create-gator-app/src/lib/helpers/validators.ts
@@ -1,0 +1,4 @@
+export const validateProjectName = (input: string) => {
+    return /^[a-zA-Z0-9-_]+$/.test(input);
+  };
+
