+++ packages/create-gator-app/src/lib/prompts/base.ts
@@ -1,25 +1,16 @@
-import { TEMPLATES } from "./templates";
-import { PackageManager } from "./package-manager";
-import { FRAMEWORKS } from "./frameworks";
+import { FRAMEWORKS_OPTIONS } from "../choices/framework";
+import { PACKAGE_MANAGER_OPTIONS } from "../choices/package-manager";
+import { TEMPLATE_OPTIONS } from "../choices/templates";
+import { validateProjectName } from "../helpers/validators";
 
-const packageManagerChoices = (
-  Object.keys({ npm: null, yarn: null, pnpm: null } as Record<
-    PackageManager,
-    null
-  >) as PackageManager[]
-).map((pm) => ({
-  name: pm,
-  value: pm,
-}));
-
-export const prompts = [
+export const BASE_PROMPTS = [
   {
     type: "input",
     name: "projectName",
     message: "What is your project named?",
     default: "my-gator-app",
     validate: (input: string) => {
-      if (/^[a-zA-Z0-9-_]+$/.test(input)) return true;
+      if (validateProjectName(input)) return true;
       return "Project name may only include letters, numbers, underscores and hashes.";
     },
   },
@@ -27,19 +18,19 @@ export const prompts = [
     type: "list",
     name: "framework",
     message: "Please choose a framework:",
-    choices: FRAMEWORKS,
+    choices: FRAMEWORKS_OPTIONS,
   },
   {
     type: "list",
     name: "packageManager",
     message: "Which package manager would you like to use?",
-    choices: packageManagerChoices,
+    choices: PACKAGE_MANAGER_OPTIONS,
   },
   {
     type: "list",
     name: "template",
     message: "Pick a template:",
-    choices: TEMPLATES,
+    choices: TEMPLATE_OPTIONS,
   },
   {
     type: "confirm",
