+++ packages/create-gator-app/src/lib/types/package-manager.ts
@@ -1,0 +1,1 @@
+export type PackageManager = "npm" | "yarn" | "pnpm";
